requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
aiohttp>=3.8.0
asyncio>=3.4.3

# Data processing and analysis
pyyaml>=6.0
python-dateutil>=2.8.0

# Web scraping and APIs
beautifulsoup4>=4.11.0
lxml>=4.9.0
feedparser>=6.0.10

# Google APIs
google-api-python-client>=2.70.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0

# Reddit API
praw>=7.6.0

# Google Trends
pytrends>=4.9.0

# Optional: Advanced features
# selenium>=4.8.0  # For advanced web scraping
# scrapy>=2.8.0    # Alternative scraping framework

# Development and testing (optional)
# pytest>=7.2.0
# black>=22.0.0
# flake8>=6.0.0
# mypy>=1.0.0